backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==1.26.4
openai==1.13.3
packaging==23.2
pillow==10.2.0
postgrest==0.16.1
proglog==0.1.10
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.17.0
